feat: refactor PaymentUpdateModal and related components for improved structure and validation handling
